Update predict.geno.values.assisted.R
Samples a proportion of lines from the validation set to add to the training set while also removing the same proportion from the training set. This is done to maintain the population size of the training population so that an improvement in accuracy will only be due to relationship and not confounded with pop size.